Refactor request/response
Refactor request/response with:
- Rename the `Message` trait to `Request` and remove most of its
  unused functionality. Also change the `TYPE_ID` to `u16`.
- Avoid the `Message` `trait` for the requests. Use `Deserialize`,
  `Serialize` and `Send` as the required traits for a response.
- Change the network interface for request response to use the
  standard `Result` instead of `ResponseMessage`.
- Combined the inbound and outbound errors in a single `enum`:
  `RequestError`.
- Remove the header (and CRC) from the `MessageCodec` to simplify
  request/response messages.
- Remove the concept of a default response and use an empty response
  instead.

+++ network-mock/src/hub.rs
@@ -14,13 +14,13 @@ use crate::{MockAddress, MockPeerId, ObservableHashMap};
 pub(crate) struct SenderKey {
     pub network_recipient: MockAddress,
     pub sender_peer: MockPeerId,
-    pub message_type: u64,
+    pub message_type: u16,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub(crate) struct RequestKey {
     pub recipient: MockAddress,
-    pub message_type: u64,
+    pub message_type: u16,
 }
 
 #[derive(Debug)]
